+++ main.py
@@ -2,7 +2,7 @@
 
 import CameraInput
 import RobotFunctions
-import RobotUI
+import RobotAPI
 
 import time
